+++ docs/release-notes.md
@@ -9,6 +9,12 @@ of `zizmor`.
 
 ## Next (UNRELEASED)
 
+### Enhancements 🌱
+
+* The [bot-conditions] audit now supports auto-fixes for many findings (#921)
+* The [bot-conditions] audit now produces findings on triggers other than
+  `pull_request_target` (#921)
+
 ## 1.10.0
 
 This is a **huge** new release, with multiple new features, enhancements,